comment state및 관련함수 이동 (MainSb -> FeedArticle)
