+++ src/pages/seulbiKim/Main/component/Comment.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { AiFillHeart, AiOutlineHeart } from 'react-icons/ai';
 import { CgRemove } from 'react-icons/cg';
 
-const Comment = ({ commentData, commentRemove, liked }) => {
+const Comment = ({ commentData, commentRemove }) => {
   return (
     <li>
       <div className="commentsCommented">
